# fix sampler and fix_bn bugs and provide a faster triplet