@wj1tr0y 2022-04-16
> Awesome PR!
> As you mentioned, the mask indexing is faster but may give rise to a data alignment problem in some cases.
> I think set replace to True can fix this hidden issue.
> However, the implementation of the torch. where can avoid it fundamentally.
> We need a trade-off, or someone tells us whether operation (replace=True/False for the sampler in various cases) is better for the final performance.

In my implementation, the behavior of sampler is not changed when there is sufficient sequences but the old one yeilds